requirements.txt
functions-framework
pandas
openpyxl
gspread
google-auth
google-cloud-bigquery
pyyaml
python-dotenv
